+++ pyabc/epsilon/temperature.py
@@ -19,18 +19,24 @@ class Temperature(Epsilon):
 
     Parameters
     ----------
-    schemes: Union[Callable, List[Callable]]
+    schemes: Union[Callable, List[Callable]], optional
         Temperature schemes returning proposed
         temperatures for the next time point, e.g.
         instances of :class:`pyabc.epsilon.TemperatureScheme`.
-    aggregate_fun: Callable[List[float], float]
+    aggregate_fun: Callable[List[float], float], optional
         The function to aggregate the schemes by, of the form
         ``Callable[List[float], float]``.
         Defaults to taking the minimum.
-    initial_temperature: float
+    initial_temperature: float, optional
         The initial temperature. If None provided, an AcceptanceRateScheme
         is used.
-    maximum_nr_populations: int
+    enforce_exact_final_temperature: bool, optional
+        Whether to force the final temperature (if max_nr_populations < inf)
+        to be 1.0, giving exact inference.
+
+    Properties
+    ----------
+    max_nr_populations: int
         The maximum number of iterations as passed to ABCSMC.
         May be inf, but not all schemes can handle that (and will complain).
     temperatures: Dict[int, float]
@@ -41,10 +47,8 @@ class Temperature(Epsilon):
             self,
             schemes: Union[Callable, List[Callable]] = None,
             aggregate_fun: Callable[[List[float]], float] = None,
-            initial_temperature: float = None):
-        if schemes is None:
-            # this combination proved rather stable
-            schemes = [AcceptanceRateScheme(), ExponentialDecayScheme()]
+            initial_temperature: float = None,
+            enforce_exact_final_temperature: bool = True):
         self.schemes = schemes
 
         if aggregate_fun is None:
@@ -55,6 +59,8 @@ class Temperature(Epsilon):
         if initial_temperature is None:
             initial_temperature = AcceptanceRateScheme()
         self.initial_temperature = initial_temperature
+
+        self.enforce_exact_final_temperature = enforce_exact_final_temperature
 
         # to be filled later
         self.max_nr_populations = None
@@ -67,6 +73,15 @@ class Temperature(Epsilon):
                    max_nr_populations: int,
                    acceptor_config: dict):
         self.max_nr_populations = max_nr_populations
+
+        # set default schemes
+        if self.schemes is None:
+            # this combination proved rather stable
+            acc_rate_scheme = AcceptanceRateScheme()
+            decay_scheme = (
+                ExpDecayFixedIterScheme() if np.isfinite(max_nr_populations)
+                else ExpDecayFixedRatioScheme())
+            self.schemes = [acc_rate_scheme, decay_scheme]
 
         # set initial temperature for time t
         self._update(t, get_weighted_distances, get_all_records,
@@ -110,7 +125,8 @@ class Temperature(Epsilon):
             acceptance_rate=acceptance_rate,
         )
 
-        if t >= self.max_nr_populations - 1:
+        if t >= self.max_nr_populations - 1 \
+                and self.enforce_exact_final_temperature:
             # t is last time
             temperature = 1.0
         elif not self.temperatures:  # need an initial value
@@ -207,12 +223,19 @@ class AcceptanceRateScheme(TemperatureScheme):
 
     Parameters
     ----------
-    target_rate: float
+    target_rate: float, optional
         The target acceptance rate to match.
+    min_rate: float, optional
+        The minimum rate below which not to apply the acceptance step scheme
+        any more. Setting this to a value of e.g. 0.05 can make sense
+        1) because it may be unlikely that the acceptance rate scheme will
+        propose a useful temperature at such low acceptance levels, and
+        2) to avoid uneccessary computations.
     """
 
-    def __init__(self, target_rate: float = 0.3):
+    def __init__(self, target_rate: float = 0.3, min_rate: float = None):
         self.target_rate = target_rate
+        self.min_rate = min_rate
 
     def configure_sampler(self, sampler: Sampler):
         sampler.sample_factory.record_rejected = True
@@ -226,6 +249,10 @@ class AcceptanceRateScheme(TemperatureScheme):
                  kernel_scale: str,
                  prev_temperature: float,
                  acceptance_rate: float):
+        # check minimum rate
+        if self.min_rate is not None and acceptance_rate < self.min_rate:
+            return np.inf
+
         # execute function (expensive if in calibration)
         records = get_all_records()
         # convert to dataframe for easier extraction
@@ -237,45 +264,66 @@ class AcceptanceRateScheme(TemperatureScheme):
         # acceptance kernel likelihoods
         pds = np.array(records['distance'], dtype=float)
 
-        # compute rescaled posterior densities
-        if kernel_scale == SCALE_LIN:
-            acc_probs = pds / pdf_norm
-        else:  # kernel_scale == SCALE_LOG
-            acc_probs = np.exp(pds - pdf_norm)
-
-        # to acceptance probabilities
-        acc_probs = np.minimum(acc_probs, 1.0)
-
         # compute importance weights
         weights = t_pd / t_pd_prev
         # len would suffice, but maybe rather not rely on things to be normed
         weights /= sum(weights)
 
-        # objective function which we wish to find a root for
-        def obj(beta):
-            val = np.sum(weights * acc_probs**beta) - self.target_rate
-            return val
+        temperature = match_acceptance_rate(
+            weights, pds, pdf_norm, kernel_scale, self.target_rate)
 
-        if obj(1) > 0:
-            # function is monotonically decreasing
-            # smallest possible value already > 0
-            beta_opt = 1.0
-            # it is obj(0) >= 0 always
-        else:
-            # perform binary search
-            # TODO: check out more efficient optimization approach
-            # TODO: When the weigts are low, there are numeric problems close
-            # to 1 (especially in the first iteration)
-            beta_opt = sp.optimize.bisect(
-                obj, 0, 1, maxiter=100000)
-        # temperature is inverse beta
-        temperature = 1.0 / beta_opt
         return temperature
 
 
-class ExponentialDecayScheme(TemperatureScheme):
+def match_acceptance_rate(
+        weights, pds, pdf_norm, kernel_scale, target_rate):
     """
-    If `max_nr_populations` is finite:
+    For large temperature, changes become effective on an exponential scale,
+    thus we optimize the logarithm of the inverse temperature beta.
+
+    For a temperature close to 1, subtler changes are neccesary, however here
+    the logarhtm is nearly linear anyway.
+    """
+    # objective function which we wish to find a root for
+    def obj(b):
+        beta = np.exp(b)
+
+        # compute rescaled posterior densities
+        if kernel_scale == SCALE_LIN:
+            acc_probs = (pds / pdf_norm) ** beta
+        else:  # kernel_scale == SCALE_LOG
+            acc_probs = np.exp((pds - pdf_norm) * beta)
+
+        # to acceptance probabilities to be sure
+        acc_probs = np.minimum(acc_probs, 1.0)
+
+        # objective function
+        val = np.sum(weights * acc_probs) - target_rate
+        return val
+
+    # TODO the lower boundary min_b is somewhat arbitrary
+    min_b = -100
+    if obj(0) > 0:
+        # function is monotonically decreasing
+        # smallest possible value already > 0
+        b_opt = 0
+    elif obj(min_b) < 0:
+        # it is obj(-inf) > 0 always
+        logger.info("AcceptanceRateScheme: Numerics limit temperature.")
+        b_opt = min_b
+    else:
+        # perform binary search
+        b_opt = sp.optimize.bisect(obj, min_b, 0, maxiter=100000)
+
+    beta_opt = np.exp(b_opt)
+
+    temperature = 1. / beta_opt
+    return temperature
+
+
+class ExpDecayFixedIterScheme(TemperatureScheme):
+    """
+    The next temperature is set as
 
     .. math::
         T_j = T_{max}^{(n-j)/n}
@@ -288,13 +336,6 @@ class ExponentialDecayScheme(TemperatureScheme):
 
     This ensures that a temperature of 1.0 is reached after exactly the
     remaining number of steps.
-
-    If `max_nr_populations` is infinite, the next temperature is
-
-    .. math::
-        T_j = \\alpha \\cdot T_{j-1},
-
-    where by default alpha=0.25.
 
     So, in both cases the sequence of temperatures follows an exponential
     decay, also known as a geometric progression, or a linear progression
@@ -312,8 +353,8 @@ class ExponentialDecayScheme(TemperatureScheme):
         is infinite.
     """
 
-    def __init__(self, alpha: float = 0.25):
-        self.alpha = alpha
+    def __init__(self):
+        pass
 
     def __call__(self,
                  t: int,
@@ -324,6 +365,12 @@ class ExponentialDecayScheme(TemperatureScheme):
                  kernel_scale: str,
                  prev_temperature: float,
                  acceptance_rate: float):
+        # needs a finite number of iterations
+        if max_nr_populations == np.inf:
+            raise ValueError(
+                "The ExpDecayFixedIterScheme requires a finite "
+                "`max_nr_populations`.")
+
         # needs a starting temperature
         # if not available, return infinite temperature
         if prev_temperature is None:
@@ -331,11 +378,6 @@ class ExponentialDecayScheme(TemperatureScheme):
 
         # base temperature
         temp_base = prev_temperature
-
-        if max_nr_populations == np.inf:
-            # just decrease by a factor of alpha each round
-            temperature = self.alpha * temp_base
-            return temperature
 
         # how many steps left?
         t_to_go = max_nr_populations - t
@@ -346,7 +388,76 @@ class ExponentialDecayScheme(TemperatureScheme):
         return temperature
 
 
-class PolynomialDecayScheme(TemperatureScheme):
+class ExpDecayFixedRatioScheme(TemperatureScheme):
+    """
+    The next temperature is chosen as
+
+    .. math::
+        T_j = \\alpha \\cdot T_{j-1}.
+
+    Like the :class:`pyabc.epsilon.ExpDecayFixedIterScheme`,
+    this yields a geometric progression, however with a fixed ratio,
+    irrespective of the number of iterations. If a finite number of
+    iterations is specified in ABCSMC, there is no influence on the final
+    jump to a temperature of 1.0.
+
+    This is quite similar to the :class:`pyabc.epsilon.DalyScheme`, although
+    simpler in implementation. The alpha value here corresponds to a value of
+    1 - alpha there.
+
+    Parameters
+    ----------
+    alpha: float, optional
+        The ratio of subsequent temperatures.
+    min_rate: float, optional
+        A minimum acceptance rate. If this rate has been violated in the
+        previous iteration, the alpha value is increased.
+    max_rate: float, optional
+        Maximum rate to not be exceeded, otherwise the alpha value is
+        decreased.
+    """
+    def __init__(self, alpha: float = 0.5,
+                 min_rate: float = 1e-4, max_rate: float = 0.5):
+        self.alpha = alpha
+        self.min_rate = min_rate
+        self.max_rate = max_rate
+        self.alphas = {}
+
+    def __call__(self,
+                 t: int,
+                 get_weighted_distances: Callable[[], pd.DataFrame],
+                 get_all_records: Callable[[], List[dict]],
+                 max_nr_populations: int,
+                 pdf_norm: float,
+                 kernel_scale: str,
+                 prev_temperature: float,
+                 acceptance_rate: float):
+        if prev_temperature is None:
+            return np.inf
+
+        # previous alpha
+        alpha = self.alphas.get(t-1, self.alpha)
+
+        # check if acceptance rate criterion violated
+        if acceptance_rate > self.max_rate and t > 1:
+            logger.debug("ExpDecayFixedRatioScheme: "
+                         "Reacting to high acceptance rate.")
+            alpha = max(alpha / 2, alpha - (1 - alpha) * 2)
+        if acceptance_rate < self.min_rate:
+            logger.debug("ExpDecayFixedRatioScheme: "
+                         "Reacting to low acceptance rate.")
+            # increase alpha
+            alpha = alpha + (1 - alpha) / 2
+        # record
+        self.alphas[t] = alpha
+
+        # reduce temperature
+        temperature = self.alphas[t] * prev_temperature
+
+        return temperature
+
+
+class PolynomialDecayFixedIterScheme(TemperatureScheme):
     """
     Compute next temperature as pre-last entry in
 
@@ -355,8 +466,9 @@ class PolynomialDecayScheme(TemperatureScheme):
 
     Requires finite `max_nr_populations`.
 
-    Note that this is similar to the `scheme_exponential_decay`, which is
-    indeed the limit for `temp_decay_exponent -> infinity`. For smaller
+    Note that this is similar to the
+    :class:`pyabc.epsilon.ExpDecayFixedIterScheme`, which is
+    indeed the limit for `exponent -> infinity`. For smaller
     exponent, the sequence makes larger steps for low temperatures. This
     can be useful in cases, where lower temperatures (which are usually
     more expensive) can be traversed in few larger steps, however also
@@ -417,10 +529,21 @@ class DalyScheme(TemperatureScheme):
     too low acceptance rates, such that this can only be done ex-posteriori
     here.
 
-    .. [#daly2017] Daly Aidan C., Cooper Jonathan, Gavaghan David J. ,
+    Parameters
+    ----------
+    alpha: float, optional
+        The ratio by which to decrease the temperature value. More
+        specifically, the next temperature is given as
+        `(1-alpha) * temperature`.
+    min_rate: float, optional
+        A minimum acceptance rate. If this rate has been violated in the
+        previous iteration, the alpha value is decreased.
+
+
+    .. [#daly2017] Daly Aidan C., Cooper Jonathan, Gavaghan David J.,
             and Holmes Chris. "Comparing two sequential Monte Carlo samplers
             for exact and approximate Bayesian inference on biological
-            models". Journal of The Royal Society Interface, 2017
+            models". Journal of The Royal Society Interface, 2017.
     """
 
     def __init__(self, alpha: float = 0.5, min_rate: float = 1e-4):