+++ pyabc/visualization/histogram.py
@@ -2,11 +2,13 @@ import matplotlib.pyplot as plt
 import pandas as pd
 
 from ..storage import History
+from .util import format_plot_matrix
 
 
 def plot_histogram_1d(
         history: History, x: str, m: int = 0, t: int = None,
-        xmin=None, xmax=None, ax=None, size=None, refval=None, **kwargs):
+        xmin=None, xmax=None, ax=None, size=None, refval=None,
+        refval_color='C1', **kwargs):
     """
     Plot 1d histogram of parameter samples.
 
@@ -29,6 +31,8 @@ def plot_histogram_1d(
         Size of the plot in inches.
     refval: dict, optional (default = None)
         A reference value for x, to be highlighted in the plot.
+    refval_color: str, optional
+        Color to use for the reference value.
 
     Returns
     -------
@@ -38,13 +42,14 @@ def plot_histogram_1d(
     df, w = history.get_distribution(m=m, t=t)
 
     return plot_histogram_1d_lowlevel(
-        df, w, x, xmin, xmax, ax=ax, size=size, refval=refval, **kwargs)
+        df, w, x, xmin, xmax, ax=ax, size=size, refval=refval,
+        refval_color=refval_color, **kwargs)
 
 
 def plot_histogram_1d_lowlevel(
         df: pd.DataFrame, w: pd.DataFrame,
         x: str, xmin=None, xmax=None, ax=None, size=None, refval=None,
-        **kwargs):
+        refval_color='C1', **kwargs):
     """
     Lowlevel interface for plot_histogram_1d (see there for the remaining
     parameters).
@@ -66,7 +71,7 @@ def plot_histogram_1d_lowlevel(
     else:
         range_ = None
     if refval is not None:
-        ax.axvline(refval[x], color='C1', linestyle='dashed')
+        ax.axvline(refval[x], color=refval_color, linestyle='dotted')
 
     # plot
     ax.hist(x=df[x], range=range_, weights=w, density=True, **kwargs)
@@ -82,7 +87,7 @@ def plot_histogram_1d_lowlevel(
 def plot_histogram_2d(
         history: History, x: str, y: str, m: int = 0, t: int = None,
         xmin=None, xmax=None, ymin=None, ymax=None, ax=None, size=None,
-        refval=None, **kwargs):
+        refval=None, refval_color='C1', **kwargs):
     """
     Plot 2d histogram of parameter pair samples.
 
@@ -105,6 +110,8 @@ def plot_histogram_2d(
         Size of the plot in inches.
     refval: dict, optional (default = None)
         Reference values for x and y, to be highlighted in the plot.
+    refval_color: str, optional
+        Color to use for the reference value.
 
     Returns
     -------
@@ -115,13 +122,13 @@ def plot_histogram_2d(
 
     return plot_histogram_2d_lowlevel(
         df, w, x, y, xmin, xmax, ymin, ymax, ax=ax, size=size, refval=refval,
-        **kwargs)
+        refval_color='C1', **kwargs)
 
 
 def plot_histogram_2d_lowlevel(
         df: pd.DataFrame, w: pd.DataFrame,
         x, y, xmin=None, xmax=None, ymin=None, ymax=None, ax=None,
-        size=None, refval=None, **kwargs):
+        size=None, refval=None, refval_color='C1', **kwargs):
     """
     Lowlevel interface for plot_histogram_2d (see there for the remaining
     parameters).
@@ -151,7 +158,7 @@ def plot_histogram_2d_lowlevel(
     ax.hist2d(x=df[x], y=df[y], range=range_, weights=w, density=True,
               **kwargs)
     if refval is not None:
-        ax.scatter([refval[x]], [refval[y]], color='C1')
+        ax.scatter([refval[x]], [refval[y]], color=refval_color)
     ax.set_xlabel(x)
     ax.set_ylabel(y)
 
@@ -164,7 +171,7 @@ def plot_histogram_2d_lowlevel(
 
 def plot_histogram_matrix(
         history: History, m: int = 0, t: int = None, size=None, refval=None,
-        **kwargs):
+        refval_color='C1', **kwargs):
     """
     Plot matrix of 1d and 2d histograms over all parameters.
 
@@ -181,6 +188,8 @@ def plot_histogram_matrix(
         Size of the plot in inches.
     refval: dict, optional (default = None)
         Reference parameter values, to be highlighted in the plot.
+    refval_color: str, optional
+        Color to use for the reference value.
 
     Returns
     -------
@@ -190,11 +199,13 @@ def plot_histogram_matrix(
     """
     df, w = history.get_distribution(m=m, t=t)
 
-    return plot_histogram_matrix_lowlevel(df, w, size, refval, **kwargs)
+    return plot_histogram_matrix_lowlevel(
+        df, w, size, refval, refval_color, **kwargs)
 
 
 def plot_histogram_matrix_lowlevel(
-        df: pd.DataFrame, w: pd.DataFrame, size=None, refval=None, **kwargs):
+        df: pd.DataFrame, w: pd.DataFrame, size=None, refval=None,
+        refval_color='C1', **kwargs):
     """
     Lowlevel interface for plot_histogram_matrix (see there for the remaining
     parameters).
@@ -217,7 +228,7 @@ def plot_histogram_matrix_lowlevel(
     def scatter(x, y, ax, refval=None):
         ax.scatter(x, y, color="k")
         if refval is not None:
-            ax.scatter([refval[x.name]], [refval[y.name]], color='C1')
+            ax.scatter([refval[x.name]], [refval[y.name]], color=refval_color)
 
     # fill all subplots
     for i in range(0, n_par):
@@ -226,8 +237,9 @@ def plot_histogram_matrix_lowlevel(
 
         # diagonal
         ax = arr_ax[i, i]
-        plot_histogram_1d_lowlevel(df, w, y_name, ax=ax, refval=refval,
-                                   **kwargs)
+        plot_histogram_1d_lowlevel(
+            df, w, y_name, ax=ax, refval=refval, refval_color=refval_color,
+            **kwargs)
 
         for j in range(0, i):
             x_name = par_names[j]
@@ -235,15 +247,16 @@ def plot_histogram_matrix_lowlevel(
 
             # lower
             ax = arr_ax[i, j]
-            plot_histogram_2d_lowlevel(df, w, x_name, y_name, ax=ax,
-                                       refval=refval, **kwargs)
+            plot_histogram_2d_lowlevel(
+                df, w, x_name, y_name, ax=ax, refval=refval,
+                refval_color=refval_color, **kwargs)
 
             # upper
             ax = arr_ax[j, i]
             scatter(y, x, ax, refval=refval)
 
     # format
-    _format_histogram_matrix(arr_ax, par_names)
+    format_plot_matrix(arr_ax, par_names)
 
     # set size
     if size is not None:
@@ -252,29 +265,3 @@ def plot_histogram_matrix_lowlevel(
     fig.tight_layout()
 
     return arr_ax
-
-
-def _format_histogram_matrix(arr_ax, par_names):
-    """
-    Apply some post-formatting to tidy up the plot.
-    """
-    n_par = len(par_names)
-
-    for i in range(0, n_par):
-        for j in range(0, n_par):
-            # clear labels
-            arr_ax[i, j].set_xlabel("")
-            arr_ax[i, j].set_ylabel("")
-
-            # clear legends
-            arr_ax[i, j].legend = None
-
-            # remove spines
-            arr_ax[i, j].spines['right'].set_visible(False)
-            arr_ax[i, j].spines['top'].set_visible(False)
-
-    # set left-most and bottom-most labels to parameter names
-    for ax, label in zip(arr_ax[-1, :], par_names):
-        ax.set_xlabel(label)
-    for ax, label in zip(arr_ax[:, 0], par_names):
-        ax.set_ylabel(label)