+++ test/test_epsilon.py
@@ -53,47 +53,45 @@ def test_medianepsilon():
 
 
 def test_temperature():
-    weighted_distances = pd.DataFrame({
-        'distance': [1, 2, 3, 4],
-        'w': [2, 1, 1, 0]})
-    all_records = [
-        dict(distance=np.random.randn(),
-             transition_pd_prev=np.random.randn(),
-             transition_pd=np.random.randn(),
-             accepted=True if np.random.random() > 0.5 else False)
-        for _ in range(20)]
     acceptor_config = {
         'pdf_norm': 5,
         'kernel_scale': pyabc.distance.SCALE_LOG}
     nr_pop = 3
     eps = pyabc.Temperature(initial_temperature=42)
-    eps.initialize(0, lambda: weighted_distances, lambda: all_records,
+    eps.initialize(0, get_weighted_distances, get_all_records,
                    nr_pop, acceptor_config)
 
     # check if initial value is respected
     assert eps(0) == 42
 
-    eps.update(1, lambda: weighted_distances, lambda: all_records,
+    eps.update(1, get_weighted_distances, get_all_records,
                0.4, acceptor_config)
     assert eps(1) < 42
 
     # last time
-    eps.update(2, lambda: weighted_distances, lambda: all_records,
+    eps.update(2, get_weighted_distances, get_all_records,
                0.2, acceptor_config)
     assert eps(2) == 1
 
 
-scheme_args = dict(
-    get_weighted_distances=lambda: pd.DataFrame({
-        'distance': [4, 5, 6, 1],
-        'w': [0.25, 0.5, 0.10, 0.15]
-    }),
-    get_all_records=lambda: [
+def get_weighted_distances():
+    return pd.DataFrame({
+        'distance': [1, 2, 3, 4],
+        'w': [2, 1, 1, 0]})
+
+
+def get_all_records():
+    return [
         dict(distance=np.random.randn(),
              transition_pd_prev=np.random.randn(),
              transition_pd=np.random.randn(),
              accepted=True if np.random.random() > 0.5 else False)
-        for _ in range(20)],
+        for _ in range(20)]
+
+
+scheme_args = dict(
+    get_weighted_distances=get_weighted_distances,
+    get_all_records=get_all_records,
     max_nr_populations=3,
     pdf_norm=10,
     kernel_scale=pyabc.distance.SCALE_LOG,
@@ -104,8 +102,9 @@ scheme_args = dict(
 def test_scheme_basic():
     schemes = [
         pyabc.AcceptanceRateScheme(),
-        pyabc.ExponentialDecayScheme(),
-        pyabc.PolynomialDecayScheme(),
+        pyabc.ExpDecayFixedIterScheme(),
+        pyabc.ExpDecayFixedRatioScheme(),
+        pyabc.PolynomialDecayFixedIterScheme(),
         pyabc.DalyScheme(),
         pyabc.FrielPettittScheme(),
         pyabc.EssScheme()
@@ -124,14 +123,15 @@ def test_scheme_acceptancerate():
     # high acceptance probabilities
     _scheme_args = copy.deepcopy(scheme_args)
     # change normalization s.t. most have 1.0 acceptance rate
-    records = pd.DataFrame(_scheme_args['get_all_records']())
-    _scheme_args['pdf_norm'] = min(records['distance'])
+    records = _scheme_args['get_all_records']()
+    _scheme_args['pdf_norm'] = min(pd.DataFrame(records)['distance'])
+    _scheme_args['get_all_records'] = lambda: records
     temp = scheme(t=0, **_scheme_args)
     assert temp == 1.0
 
 
 def test_scheme_exponentialdecay():
-    scheme = pyabc.ExponentialDecayScheme()
+    scheme = pyabc.ExpDecayFixedIterScheme()
 
     # check no base temperature
     _scheme_args = copy.deepcopy(scheme_args)